Aichat: split aichatLab redux slice from aichat slice
Chat session state (messages, staged files, selection context) stays in
the aichat slice. Model customization state (initial/current/saved
customizations, visibilities, save flow, view mode) moves to a new
aichatLab slice under apps/src/aichat/aichatLab/redux, along with its
thunks, selectors, and utils. Components and tests consume the lab
slice via @cdo/apps/aichat/aichatLab/redux.
