+++ apps/src/aichat/aichatLab/redux/thunks/initializeAiCustomizations.ts
@@ -1,16 +1,19 @@
+import {sendAnalytics} from '@cdo/apps/aichat/redux';
 import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 import {getTypedKeys} from '@cdo/apps/types/utils';
 import {AppDispatch} from '@cdo/apps/util/reduxHooks';
 
-import {AiCustomizations, LevelAichatSettings, Visibility} from '../../types';
+import {
+  AiCustomizations,
+  LevelAichatSettings,
+  Visibility,
+} from '../../../types';
 import {
   DEFAULT_VISIBILITIES,
   EMPTY_AI_CUSTOMIZATIONS,
-} from '../../views/modelCustomization/constants';
-import {validateModelId} from '../../views/modelCustomization/utils';
+} from '../../../views/modelCustomization/constants';
+import {validateModelId} from '../../../views/modelCustomization/utils';
 import {setInitialConfiguration} from '../slice';
-
-import {sendAnalytics} from './sendAnalytics';
 
 /**
  * Initialize AI customizations for the level by reconciling student customizations
