+++ apps/src/aichat/aichatLab/redux/thunks/onSaveNoop.ts
@@ -1,7 +1,7 @@
 import {RootState} from '@cdo/apps/types/redux';
 import {AppDispatch} from '@cdo/apps/util/reduxHooks';
 
-import {ViewMode} from '../../types';
+import {ViewMode} from '../../../types';
 import {endSave, setViewMode} from '../slice';
 
 // Thunk called when a save no-ops (there are no changes to save)
@@ -9,7 +9,7 @@ export const onSaveNoop =
   () => (dispatch: AppDispatch, getState: () => RootState) => {
     // Even if no changes were saved, go to the presentation page if the user tried to publish
     // a model card.
-    if (getState().aichat.currentSaveType === 'publishModelCard') {
+    if (getState().aichatLab.currentSaveType === 'publishModelCard') {
       dispatch(setViewMode(ViewMode.PRESENTATION));
     }
     dispatch(endSave());
