+++ apps/src/aichat/redux/selectors/index.ts
@@ -5,34 +5,6 @@ import type {RootState} from '@cdo/apps/types/redux';
 import {AiInteractionStatus} from '@cdo/generated-scripts/sharedConstants';
 
 import {isChatMessage} from '../../types';
-import {
-  allFieldsHidden,
-  anyFieldsChanged,
-  findChangedProperties,
-  hasFilledOutModelCard,
-} from '../utils';
-
-export const selectHasFilledOutModelCard = createSelector(
-  (state: RootState) => state.aichat.currentAiCustomizations.modelCardInfo,
-  hasFilledOutModelCard
-);
-
-export const selectAllFieldsHidden = createSelector(
-  (state: RootState) => state.aichat.fieldVisibilities,
-  allFieldsHidden
-);
-
-export const selectCurrentCustomizationsMatchInitial = createSelector(
-  (state: RootState) => state.aichat.initialAiCustomizations,
-  (state: RootState) => state.aichat.currentAiCustomizations,
-  anyFieldsChanged
-);
-
-export const selectSavedCustomizationsMatchInitial = createSelector(
-  (state: RootState) => state.aichat.initialAiCustomizations,
-  (state: RootState) => state.aichat.savedAiCustomizations,
-  anyFieldsChanged
-);
 
 export const selectIsWaitingForChatResponse = (state: RootState) => {
   let lastChatEvent;
@@ -59,9 +31,3 @@ export const selectAllVisibleMessages = createSelector(
     ...chatEventsCurrent,
   ]
 );
-
-export const selectHavePropertiesChanged = (state: RootState) =>
-  findChangedProperties(
-    state.aichat.savedAiCustomizations,
-    state.aichat.currentAiCustomizations
-  ).length > 0;
