+++ apps/src/aichat/redux/state.ts
@@ -1,14 +1,8 @@
-import {ModalTypes} from '../constants';
 import {
-  AiCustomizations,
   AiChatClientType,
   ChatAsset,
   ChatEvent,
-  FieldVisibilities,
-  SaveError,
-  SaveType,
   ServerChatEvent,
-  ViewMode,
   WorkspaceTeacherViewTab,
   UserAddedSelectionContext,
 } from '../types';
@@ -22,17 +16,6 @@ export interface AichatState {
   chatEventsCurrent: ChatEvent[];
   // Student events viewed by a teacher user in chat workspace. Always fetched from the server.
   studentChatHistory: ServerChatEvent[];
-  // Denotes whether we should show the warning or teacher onboarding modal
-  showModalType: ModalTypes | undefined;
-  initialAiCustomizations: AiCustomizations;
-  currentAiCustomizations: AiCustomizations;
-  savedAiCustomizations: AiCustomizations;
-  fieldVisibilities: FieldVisibilities;
-  viewMode: ViewMode;
-  // If a save is currently in progress
-  saveInProgress: boolean;
-  // The type of save action being performed (customization update, publish, model card save, etc).
-  currentSaveType: SaveType | undefined;
   // List of files that have been staged for upload to the model.
   stagedFiles: {
     key: string;
@@ -46,18 +29,6 @@ export interface AichatState {
     | 'sizeLimitExceeded'
     | 'imageFileFlagged'
     | undefined;
-  // If the user has a sent a message on this level
-  hasSentMessage: boolean;
-  // If initial customizations have been set on this level
-  hasSetInitialCustomizations: boolean;
-  // If the user has updated customizations on this level
-  hasUpdatedCustomizations: boolean;
-  // Error message to display if a save fails
-  saveError: SaveError | undefined;
-  // If the model customizations were just reset to the default level values.
-  showResetMessage: boolean;
-  // If the user had previously selected a model that is no longer available.
-  showUnsupportedModelMessage: boolean;
   // The tab selected when a teacher is viewing a student's chat history.
   chatWorkspaceSelectedTab: WorkspaceTeacherViewTab | null;
   userAddedSelectionContext: UserAddedSelectionContext;
