+++ apps/src/aichat/views/modelCustomization/CompareModelsDialog.tsx
@@ -16,7 +16,7 @@ const CompareModelsDialog: React.FunctionComponent<{
 }> = ({onClose, availableModels}) => {
   const selectedModelId = useAppSelector(
     state =>
-      state.aichat.currentAiCustomizations.selectedModelId ||
+      state.aichatLab.currentAiCustomizations.selectedModelId ||
       availableModels[0].id
   );
   const chosenModelLeft = selectedModelId;
