+++ apps/src/aichat/views/modelCustomization/RetrievalCustomization.tsx
@@ -4,7 +4,7 @@ import {useSelector} from 'react-redux';
 import {isReadOnlyWorkspace} from '@cdo/apps/lab2/redux/lab2ReduxSelectors';
 import {useAppSelector, useAppDispatch} from '@cdo/apps/util/reduxHooks';
 
-import {setAiCustomizationProperty} from '../../redux';
+import {setAiCustomizationProperty} from '../../aichatLab/redux';
 
 import MultiInputCustomization from './MultiInputCustomization';
 import SaveChangesAlerts from './SaveChangesAlerts';
@@ -16,10 +16,10 @@ import modelCustomizationStyles from '../model-customization-workspace.module.sc
 const RetrievalCustomization: React.FunctionComponent = () => {
   const dispatch = useAppDispatch();
   const visibility = useAppSelector(
-    state => state.aichat.fieldVisibilities.retrievalContexts
+    state => state.aichatLab.fieldVisibilities.retrievalContexts
   );
   const {retrievalContexts} = useAppSelector(
-    state => state.aichat.currentAiCustomizations
+    state => state.aichatLab.currentAiCustomizations
   );
 
   const isReadOnly = useSelector(isReadOnlyWorkspace) || isDisabled(visibility);
