+++ apps/src/aichat/views/modelCustomization/UpdateButton.tsx
@@ -4,7 +4,10 @@ import React, {useCallback} from 'react';
 
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
 
-import {selectHavePropertiesChanged, updateAiCustomization} from '../../redux';
+import {
+  selectHavePropertiesChanged,
+  updateAiCustomization,
+} from '../../aichatLab/redux';
 
 import styles from '../model-customization-workspace.module.scss';
 
@@ -20,8 +23,12 @@ const UpdateButton: React.FunctionComponent<UpdateButtonProps> = ({
     () => dispatch(updateAiCustomization()),
     [dispatch]
   );
-  const saveInProgress = useAppSelector(state => state.aichat.saveInProgress);
-  const currentSaveType = useAppSelector(state => state.aichat.currentSaveType);
+  const saveInProgress = useAppSelector(
+    state => state.aichatLab.saveInProgress
+  );
+  const currentSaveType = useAppSelector(
+    state => state.aichatLab.currentSaveType
+  );
   const havePropertiesChanged = useAppSelector(selectHavePropertiesChanged);
 
   return (
