+++ apps/test/unit/aichat/utilsTest.ts
@@ -1,4 +1,4 @@
-import * as utils from '@cdo/apps/aichat/redux/utils';
+import * as utils from '@cdo/apps/aichat/aichatLab/redux/utils';
 import {AiCustomizations} from '@cdo/apps/aichat/types';
 import {EMPTY_MODEL_CARD_INFO} from '@cdo/apps/aichat/views/modelCustomization/constants';
 import {AiChatModelIds} from '@cdo/generated-scripts/sharedConstants';
